backend/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
alembic==1.16.1
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==24.3.0
babel==2.17.0
bcrypt==4.1.2
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.15
dnspython==2.6.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
fonttools==4.58.5
food-volume-estimation==0.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.5.1
fuzzywuzzy==0.18.0
gast==0.6.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.75.1
h11==0.16.0
h5py==3.15.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
ipydatawidgets==4.3.5
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
jwt==1.4.0
keras==3.11.3
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lightning==2.5.2
lightning-utilities==0.14.3
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.1.0
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.2.2
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.2.1
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyGithub==2.5.0
Pygments==2.19.2
PyJWT==2.8.0
pymongo==4.6.2
PyNaCl==1.5.0
pyntcloud==0.1.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pythreejs==2.1.1
pytorch-lightning==2.5.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.2.0
rpds-py==0.26.0
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
Send2Trash==1.8.3
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.10.4
tinycss2==1.4.0
torch==2.7.1
torchaudio==2.7.1
torchmetrics==1.7.4
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
types-python-dateutil==2.9.0.20250708
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.0
yarl==1.18.3
